refactor: place each component in own directory
