+++ src/SearchPage/Main/index.js
@@ -1,0 +1,10 @@
+import React from "react";
+import { GoUp, Main } from "./styled";
+import Card from "./Card";
+
+export default () => (
+  <Main>
+    <GoUp>Наверх</GoUp>
+    <Card />
+  </Main>
+);
