+++ src/SearchPage/UI/index.js
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+export const Button = styled.button`
+  font-family: "Roboto", sans-serif;
+  font-size: 14px;
+  line-height: 16px;
+  font-weight: 500;
+  color: #fff;
+  text-transform: uppercase;
+  border: 2px solid #fff;
+  border-radius: 100px;
+  text-align: center;
+`;
